+++ release-notes.md
@@ -606,7 +606,7 @@ In this version, we focused on security upgrades and introduced several new feat
 
 **Major changes**
 
-- The displaying layout of NFTs in the extension is updated to match the layout on https://dojak.io/
+- The displaying layout of NFTs in the extension is updated to match the layout on https://dojak.dog/
 - The default focus is set into the input field for easier keyboard input as needed
 
 **Minor changes**